==> R/01_load.qmd <==
---
title: "01_load"
format:
  html:
    embed-resources: true
editor: visual
---

### Source functions

```{r}
source("../R/99_proj_func.R")
```

### Loading libraries

```{r}
#| message: false
library(tidyverse)
library(httr)
library(jsonlite)
library(dplyr)
library(R.utils)
library(illuminaio)
```

### Load data directories

```{r}
create_dir("../_raw")
create_dir("../data")
create_dir("../results")
create_dir("../results/plots")
```

### Load files

```{r}
#| message: false
raw_dir <- "../_raw/"

download_dataset_ncbi(raw_dir)
download_data_annotation_ncbi(raw_dir)

ncbi_data <- read_ncbi_table(str_c(raw_dir, "raw_data_ncbi.txt.gz"))
annotation_data <- read_bgx_file(raw_dir, "raw_annotation.bgx.gz")

```

### Write files

The files that we have just loaded above should be written to a tsv file in our data folder. This folder contains our data frames.

```{r}
write_tsv(ncbi_data,
            file = "../data/01_ncbi_data.tsv.gz")

write_tsv(annotation_data$probes,
            file = "../data/01_ncbi_annot_probes.tsv.gz")
```
