==> R/02_clean.qmd <==
---
title: "Data Cleaning and Wrangling"
format:
  html:
    embed-resources: true
editor: visual
---

### Loading required libraries

```{r}
#| message: false
library(tidyverse)
```

### Loading the data

```{r}
#| message: false
ncbi_data <- read_tsv("../data/01_ncbi_data.tsv.gz")
ncbi_annot_probes <- read_tsv("../data/01_ncbi_annot_probes.tsv.gz")
```

### Extracting sample IDs

```{r}
# Inspecting the dataset to locate where the sample ID's are
sample_id_row <- ncbi_data[which(ncbi_data$col_1 == "ID_REF"), -1]  # Finding all places in col1 that is named ID_REF and extracts all values associated with this observation
sample_ids <- as.character(sample_id_row)                           # Transforming sample_ids to characters
```

### Extracting and filtering relevant data

```{r}
# Extracting patient characteristics data
ncbi_characteristics_dirty <- ncbi_data |> 
  as_tibble() |> 
  filter(col_1 == "!Sample_characteristics_ch1") |> 
  select(-col_1)  # Taking everything but the first column

# Extracting gene expression data
ncbi_genes_dirty <- ncbi_data |> 
  as_tibble() |> 
  filter(str_detect(col_1, "^ILMN"))  # Searching for observations starting with ILMN in col1

```

### Tidying patient characteristics data

```{r}
# Transpose and tidy the characteristics data
tidy_characteristics <- ncbi_characteristics_dirty |> 
  pivot_longer(cols = everything(),                      # all column
               names_to = "Sample",                      # all column names to sample
               values_to = "Characteristic") |>          # all values to characteristics
  separate(Characteristic, 
           into = c("Variable", "Value"),                # making a vector to split up into variable and value 
           sep = ": ",
           extra = "merge") |>                                    # separating based on 
  pivot_wider(names_from = Variable, 
              values_from = Value) |> 
  mutate(Sample_ID = sample_ids[seq_len(n())]) |>           # taking everything from sample_ids and puts it into the new Sample_ID column 
  relocate(Sample_ID)                                    # moving Sample_ID to the first column

# Removing the sample column and dropping NAs
tidy_characteristics <- tidy_characteristics |> 
  select(-Sample) |>                          # selecting everything but the "Sample" column
  mutate(across(everything(), ~ na_if(., "not available"))) |>  # replace "not available" with NA
  drop_na()                                   # dropping all NA values
  

```

### Adding gene names from annotation file

```{r}
# Ensure the column names match
ncbi_genes_dirty <- ncbi_genes_dirty |> 
  rename(Probe_Id = col_1)

# Perform a right join to add the ILMN_Gene column
placental_gene_expr <- ncbi_annot_probes |> 
  select(Probe_Id, ILMN_Gene) |> 
  right_join(ncbi_genes_dirty, by = "Probe_Id") |>
  relocate(ILMN_Gene, .after = Probe_Id)|> 
  mutate(ILMN_Gene = paste(Probe_Id, ILMN_Gene, sep = "_")) |>  # Combine Probe_Id and ILMN_Gene
    select(-1) |>
  rename(col_1 = ILMN_Gene)

```

### Tidying gene expression data

```{r}
tidy_genes <- placental_gene_expr |> 
  pivot_longer(cols = -col_1,                  # everything but col1
               names_to = "Sample",
               values_to = "Expression") |> 
  pivot_wider(names_from = col_1,             # col1 contains the genes
              values_from = Expression) |>    # all expression values
  mutate(Sample_ID = sample_ids[seq_len(n_distinct(Sample))]) |>   # adding the sample_IDs in a new column
  relocate(Sample_ID)   # moving Sample_ID to the first column



# Removing the sample column and dropping NAs
tidy_genes <- tidy_genes |> 
  ungroup() |>  # Remove any grouping by Sample
  select(-Sample) |>   # selecting everything but the "Sample" column
  mutate(across(where(is.character), ~ na_if(., "not available"))) |>  # replace "not available" with NA
  drop_na()                                   # dropping all NA values


```

### Cleaning column names

```{r}
# Standardizing column names for characteristics data 
colnames(tidy_characteristics) <- colnames(tidy_characteristics) |> 
  str_replace_all("\\s+", "_") |>           # Replacing spaces with underscores
  str_replace_all("\\.+", "_") |>           # Replacing dots with underscores
  str_replace_all("\\(|\\)", "") |>         # Removing parentheses
  str_to_lower()                            # Converting to lowercase

# Standardizing column names for gene data
colnames(tidy_genes) <- colnames(tidy_genes) |> 
  str_replace_all("\\s+", "_") |>           # Replacing spaces with underscores
  str_replace_all("\\.+", "_") |>           # Replacing dots with underscores
  str_replace_all("\\(|\\)", "") |>         # Removing parentheses
  str_to_lower()                            # Converting to lowercase


```

### Merging patient characteristics and gene expression data

```{r}
# Merge the two datasets by `sample_id`
placental_gene_expr <- tidy_characteristics |> 
  left_join(tidy_genes, by = "sample_id")  # Joining the two datasets on sample_ID

# Convert columns to numeric values
cols_not_to_convert <- c("sample_id", 
                         "status", 
                         "mode_of_delivery")
cols_to_convert <- setdiff(colnames(placental_gene_expr), 
                           cols_not_to_convert)
placental_gene_expr[cols_to_convert] <- lapply(placental_gene_expr[cols_to_convert], 
                                       function(x) as.numeric(as.character(x)))


```

### Saving the final merged data to a tsv file

```{r}
# Save the merged dataset to a tsv file
write_tsv(placental_gene_expr,
          file = "../data/02_dat_clean.tsv")
write_tsv(tidy_characteristics,
          file = "../data/02_dat_clean_characteristics.tsv")
write_tsv(tidy_genes,
          file = "../data/02_dat_clean_genes.tsv")

```
