==> R/04_describe.qmd <==
---
title: "04_describe"
format:
  html:
    embed-resources: true
editor: visual
---

### Loading libraries

```{r}
#| message: false
library(tidyverse)
library(table1)
library(patchwork)
library(ggplot2)
library(dplyr)
library(gt)
library(gridExtra)
```

### Download data

```{r}
#| message: false
placental_gene_expr <- read_tsv("../data/03_dat_augment_placental_gene_expr.tsv")
```

### Description of the Dataset

```{r}
# Number of rows and columns in the dataset 
print(paste("Number of rows in dataset:", nrow(placental_gene_expr)))
print(paste("Number of columns in dataset:",ncol(placental_gene_expr)))
```

### Creating a Summary Table of the Dataset

```{r}
#Making a summary table
summary_table_placental_gene_expr <- placental_gene_expr |> 
  group_by(status) |> 
  summarise(
    `Mothers in Cohort` = n(),
    `Age in years median (range)` = paste0(median(age_years),                                  #Calculating the median and range of age
                                            " (", min(age_years),
                                            "-", max(age_years), 
                                            ")"),
    `Maternal BMI mean (range)` = paste0(round(mean(maternal_bmi), 1),                         #Calculating the median and range of BMI
                                           " (", min(maternal_bmi), 
                                           "-", max(maternal_bmi), 
                                           ")"),
    `Parity mean` = paste0(round(mean(parity), 1)),                                            #Calculating the mean of parity
    `Gestational Age in weeks median (range)` = paste0(median(gestational_age_weeks),          #Calculating the mean and range of GA
                                                        " (", min(gestational_age_weeks),     
                                                        "-", max(gestational_age_weeks), 
                                                        ")")
  )
#Transposing the table and removing the "status" row
summary_table <- as.data.frame(t(summary_table_placental_gene_expr))
summary_table <- summary_table[-1, ]

#Remaing the columns
colnames(summary_table) <- c("non-smoker", "smoker")


#Ensuring that the final table looks right
print(summary_table)

# Saving the table af a .csv file in results folder
write.csv(summary_table, file = "../results/description_table.csv", row.names = TRUE)
```

### Box-plot over Age Filled by Smoker Status

```{r}
#| message: false
bp_age <- placental_gene_expr |>
  ggplot(aes(x = status, 
             y = age_years, 
             fill = status)) +
  geom_boxplot() +
  labs(
    x = "Status",
    y = "Age (years)",
    title = "Boxplot of Age by Smoking Status") +
  scale_fill_manual(
      values = c("non-smoker" = "#F9A1A1", 
                 "smoker" = "#63B8B8")
    ) +
  theme(
      panel.background = element_rect(fill = "white"),
      plot.background = element_rect(fill = "white"),
      panel.grid.major = element_line(color = "gray90"),
      panel.grid.minor = element_blank(),
      axis.text = element_text(size = 5, 
                               color = "black"),
      axis.title = element_text(size = 7, 
                                color = "black"),
      plot.title = element_text(size = 5, 
                                face = "bold", 
                                color = "black"))

ggsave("../results/plots/description_age_boxplot.png")

```

### Box-plot over Maternal BMI Filled by Smoker Status

```{r}
#| message: false
bp_maternal_bmi <- placental_gene_expr |>
  ggplot(aes(x = status, 
             y = maternal_bmi, 
             fill = status)) +
  geom_boxplot() +
  labs(
    x = "Status",
    y = "Maternal BMI",
    title = "Boxplot of Maternal BMI by Smoking Status") +
  scale_fill_manual(
      values = c("non-smoker" = "#F9A1A1", 
                 "smoker" = "#63B8B8")
    ) +
  theme(
      panel.background = element_rect(fill = "white"),
      plot.background = element_rect(fill = "white"),
      panel.grid.major = element_line(color = "gray90"),
      panel.grid.minor = element_blank(),
      axis.text = element_text(size = 5, 
                               color = "black"),
      axis.title = element_text(size = 7, 
                                color = "black"),
      plot.title = element_text(size = 5, 
                                face = "bold", 
                                color = "black"))

ggsave("../results/plots/description_maternal_bmi_boxplot.png")
  
```

### Box-plot over Placental Volume in cm3 filled by smoker status

```{r}
#| message: false
bp_placental_volume <- placental_gene_expr |>
  ggplot(aes(x = status, 
             y = placental_volume_cm3, 
             fill = status)) +
  geom_boxplot() +
  labs(
    x = "Status",
    y = "Placental volume in cm3",
    title = "Boxplot of Placental Volume by Smoking Status") +
  scale_fill_manual(
    values = c("non-smoker" = "#F9A1A1", 
               "smoker" = "#63B8B8")
  ) +
  theme(
    panel.background = element_rect(fill = "white"),
    plot.background = element_rect(fill = "white"),
    panel.grid.major = element_line(color = "gray90"),
    panel.grid.minor = element_blank(),
    axis.text.x = element_text(size = 5, 
                               color = "black"),
    axis.text.y = element_text(size = 8, 
                               color = "black"),
    axis.title = element_text(size = 7, 
                              color = "black"),
    plot.title = element_text(size = 5, 
                              face = "bold", 
                              color = "black"))

ggsave("../results/plots/description_placental_volume_boxplot.png")

```

### Combined plot: Correlation Between Smoker Status and Maternal Characteristics

```{r}
#| message: false
#Combined plot with correlation between smoker status and maternal characteristics: Placental volume, maternal BMI and mothers age.
combined_plot1 <- (bp_placental_volume + theme(legend.position = "none") + 
     (bp_maternal_bmi + theme(legend.position = "none")
      )) / bp_age + 
     
     plot_annotation(
       title = "Comparison by Smoking Status","\n",
       caption = "DOI: https://doi.org/10.1016/j.placenta.2009.12.016",
        ) &
  theme(
    legend.position = "topright",
    plot.title = element_text(size = 8, 
                              hjust = 0.5, 
                              face = "bold", 
                              margin = margin(0, 0, 0, 0), 
                              color = "black"),
    plot.margin = grid::unit(c(0, 0, 0, 0), "mm")
  ) +
  plot_layout(guides = "collect", 
              heights = unit(c(1.35, 2.35), 
                             c('cm', 'cm')))
ggsave("../results/plots/description_combined_plot1.png")

#Printing the plot
combined_plot1
```
