==> R/05_analysis_1.qmd <==
---
title: "05_analysis_1"
format:
  html:
    embed-resources: true
editor: visual
---

### Packages

```{r}
#| message: false
library(ggplot2)
library(dplyr)
library(tidyverse)
library(tidyr)
```

### Load data

```{r}
#| message: false
placental_gene_expr_log2_fold <- read_tsv("../data/03_dat_augment_log2_fold.tsv")
```

### Plotting the top 15 and bottom 15

```{r}
# Select the top 15 genes with the largest and the 15 genes with the smallest
top_15_genes <- placental_gene_expr_log2_fold |>
  filter(p_value < 0.05) |> 
  arrange(desc(log2_fold_change)) |>
  slice(1:15)

bottom_15_genes <- placental_gene_expr_log2_fold |>
  filter(p_value < 0.05) |> 
  arrange(log2_fold_change) |>
  slice(1:15)

# Combine the top and bottom genes
selected_genes <- bind_rows(top_15_genes, bottom_15_genes)

# Reshape the selected_genes to long format for ggplot2
selected_genes_long <- pivot_longer(
  selected_genes, 
  cols = c('non-smoker', 'smoker'), 
  names_to = "Status", 
  values_to = "Expression"
)

# Plot
ggplot(selected_genes_long, aes(x = gene, y = Expression, fill = Status)) +
  geom_bar(stat = "identity", position = "dodge") +
  labs(
    title = "Top 15 Genes with Largest and Smallest log2_fold_change",
    x = "Gene",
    y = "Mean Expression",
    fill = "Status"
  ) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

#ggsave("../results/plots/top_botom_15_genes.png")


```

####
