==> R/06_analysis_2.qmd <==
---
title: "06_analysis_2"
format:
  html:
    embed-resources: true
editor: visual
---

### Loading libraries

```{r}
#| message: false
library(dplyr)
library(ggplot2)
library(tidyverse)
```

### Download data

```{r}
#| message: false
placental_gene_expr_log2_fold <- read_tsv("../data/03_dat_augment_log2_fold.tsv")
```

### Counting significant genes

```{r}
# Count how many p-values are under 0.05 and thereby significantly different
significant_genes_count <- placental_gene_expr_log2_fold |>
  filter(p_value < 0.05) |>
  summarise(count = n())
```

### Plot of up- and downregulated genes in smokers

```{r}
ggplot(placental_gene_expr_log2_fold, aes(x = reorder(gene, log2_fold_change), y = log2_fold_change, fill = Regulation)) +
  geom_bar(stat = "identity", width = 0.8) + 
  coord_flip() +  # Flip coordinates so the plot is more interpretable 
  scale_fill_manual(values = c("Upregulated" = "#63B8B8", "Downregulated" = "#F9A1A1")) +
  labs(title = "Upregulated vs. Downregulated Genes in Smokers",
       x = "Genes",
       y = "Log2-Fold Change") +
  theme(
      panel.background = element_rect(fill = "white"),
      plot.background = element_rect(fill = "white"),
      panel.grid.major = element_line(color = "gray90"),
      panel.grid.minor = element_blank(),
      axis.text = element_text(size = 5, 
                               color = "black"),
      axis.title = element_text(size = 7, 
                                color = "black"),
      plot.title = element_text(size = 5, 
                                face = "bold", 
                                color = "black")) +
  theme(axis.text.y = element_blank(),  # Removing y-axis gene labels
        plot.title = element_text(hjust = 0.5))

ggsave("../results/plots/upregulated_downregulated_genes.png")
```
