==> R/07_analysis_3.qmd <==
---
title: "07_analysis 3"
format:
  html:
    embed-resources: true
editor: visual
---

### Loading libraries

```{r}
#| message: false
library(tidyverse)
library(broom)
library(patchwork)
library(ggplot2)
```

### Download data

```{r}
placental_gene_expr_clean_aug <- read_tsv("../data/03_dat_augment_placental_gene_expr_clean_aug.tsv")
```

## Linear regression for genes associated with smoking

### Forest plot and Volcano plot (q value)

### Creating a table with status, gene names and log2 expression levels

```{r}
#Transforming the data to a long format to prepare it for analysis
placental_gene_expr_clean_aug_long <- placental_gene_expr_clean_aug |> 
  select(status_num, starts_with("ilmn")) |>                      #Selecting relevant columns
  pivot_longer(cols = starts_with("ilmn"),                        #Specifying which columns to be pivot
               names_to = "gene",                                 #Making a new column to store the gene names
               values_to = "expr_level"                           #Making a new column to store the corresponding expression levels in
               ) |> 
  mutate(log2_expr = log2(expr_level))                            #

placental_gene_expr_clean_aug_long <- placental_gene_expr_clean_aug_long |> 
  select(status_num,gene,log2_expr)

print(placental_gene_expr_clean_aug_long)



```

#### Creating a table with gene names and a data column containing status and log2 expression level for each gene

```{r}
placental_gene_expr_clean_aug_long_nested <- placental_gene_expr_clean_aug_long |> 
  group_by(gene) |> 
  nest() |>  
  ungroup()

placental_gene_expr_clean_aug_long_nested
```

```{r}
placental_gene_expr_clean_aug_long_nested |>
  filter(gene == "ilmn_1343291") |> 
  pull(data)
```

#### Creating a column with a model_object column

```{r}
placental_gene_expr_clean_aug_long_nested <- placental_gene_expr_clean_aug_long_nested |> 
  group_by(gene) |> 
  mutate(model_object = map(.x = data,
                   .f = ~lm(formula = log2_expr ~ status_num,
                            data = .x)))
placental_gene_expr_clean_aug_long_nested
```

```{r}
placental_gene_expr_clean_aug_long_nested |>
  
  filter(gene == "ilmn_1343293") |> 
  
  # Pull() on tibbles: This pulls out the model_object variable.
  #   Note! This is a list, because we nested!
  pull(model_object) |> 
  
  # Pluck() on lists: From the list we got from the last step,
  #   we "pluck" the first element
  pluck(1) |>
  
  # The result of pluck, is a model object,
  #   upon which we can call the tidy function
  tidy(conf.int = TRUE,
       conf.level = 0.95)
```

#### Creating a column with model_object_tidy

```{r}
placental_gene_expr_clean_aug_long_nested <- placental_gene_expr_clean_aug_long_nested |> 
  mutate(model_object_tidy = map(.x = model_object, #specifies that the function will be applied to the model_object column
                                 .f = ~tidy(x = .x, # The x argument refers to the current model object being processed from the model_object column
                                            conf.int = TRUE, #include confidence intervals 
                                            conf.level = 0.95)))
placental_gene_expr_clean_aug_long_nested
```

#### Unnest the model_object_tidy column

```{r}
placental_gene_expr_estimates <- placental_gene_expr_clean_aug_long_nested |> 
  unnest(model_object_tidy)

placental_gene_expr_estimates
```

#### Filter, Select, and Convert Data Types for Gene Expression Estimates

```{r}
placental_gene_expr_estimates <- placental_gene_expr_estimates |> 
  filter(term == "status_num") |> 
  select(gene, p.value, estimate, conf.low, conf.high) |> 
  ungroup()

cols_to_convert <- c("p.value", "estimate", "conf.low", "conf.high")
placental_gene_expr_estimates[cols_to_convert] <- lapply(placental_gene_expr_estimates[cols_to_convert],
                                            function(x) as.numeric(as.character(x)))

placental_gene_expr_estimates
```

#### Finding q-value and determining if the results are significant

```{r}
placental_gene_expr_estimates <- placental_gene_expr_estimates |> 
  mutate(q.value = p.adjust(p.value),
         is_significant = case_when(q.value <= 0.05 ~ "yes",
                                    q.value > 0.05 ~ "no"))

placental_gene_expr_estimates
```

#### Finding the genes which are significant

```{r}
placental_gene_expr_estimates |>
  filter(is_significant == "yes") |> 
  select(gene, p.value, q.value)
```

No significant q-values found. Plotting all estimates (even though they are not significant):

```{r}
placental_gene_expr_estimates |> 
  filter(is_significant == "no") |> 
  ggplot(aes(x = estimate,
             y = fct_reorder(gene, estimate),
             xmin = conf.low,
             xmax = conf.high)) +
  geom_vline(xintercept = 0) +
  geom_errorbarh() +
  geom_point() +
  theme_minimal(base_size = 16) +
  theme(plot.title = element_text(hjust = 1, size = 11),
        axis.title = element_text(size = 10),
        axis.text.y = element_blank(),  # Remove y-axis labels
        axis.text = element_text(size = 5),
        plot.caption = element_text(size = 7)) +
  labs(x = "Estimates (95%CIs)",
       y = "Genes",
       title = "Genes Associated with Smoking (not significant)",
       caption = "DOI: https://doi.org/10.1016/j.placenta.2009.12.016")

ggsave("../results/plots/genes_associated_with_smoking_all.png")

```

### Plotting of genes which are significant/not significant

```{r}
library("ggrepel")
volcono_plot_q <- placental_gene_expr_estimates |>
  mutate(lbl = case_when(
    is_significant == "yes" ~ gene,
    is_significant == "no" ~ "")) |> 
  ggplot(aes(x = estimate,
             y = -log10(p.value),
             colour = is_significant,
             label = lbl)) +
  geom_point(size = 1,
             alpha = 0.5) +
  geom_text_repel(size = 4,
                  max.overlaps = 20) +
  geom_hline(yintercept = 0) +
  theme_minimal(base_size = 10) +
  theme(
    plot.title = element_text(hjust = 1),
    plot.subtitle = element_text(hjust = 1),
    legend.position = "none",
    plot.caption = element_text(size = 6)  # Set smaller caption size here
  ) +
  labs(
    x = "Estimates",
    y = "-log10(p)",
    title = "Volcano plot for down/up regulated genes with significance q < 0.05",
    subtitle = "Significant genes highlighted with turquoise - no significant genes found",
    caption = "DOI: https://doi.org/10.1016/j.placenta.2009.12.016")

volcono_plot_q
ggsave("../results/plots/volcano_plot.png")
```

Since no significant q-values were found, we decided to look at significant p-values instead.

## Significant p-values

```{r}
# Changed from looking at significant q-value in 07_analysis to looking at significant p-values
placental_gene_expr_estimates_p <- placental_gene_expr_estimates |> 
  mutate(q.value = p.adjust(p.value),
         is_significant = case_when(p.value <= 0.05 ~ "yes", # obs p-value
                                    p.value > 0.05 ~ "no"))

placental_gene_expr_estimates_p
```

#### Sorting for genes with a significant p-value (p\<0.05)

```{r}
placental_gene_expr_estimates_p |>
  filter(is_significant == "yes") |> 
  select(gene, p.value, q.value)
```

### Forest plot for all genes with a significant p-value

```{r}
placental_gene_expr_estimates_p |> 
  filter(is_significant == "yes") |> 
  ggplot(aes(x = estimate,
             y = fct_reorder(gene, estimate),
             xmin = conf.low,
             xmax = conf.high)) +
  geom_vline(xintercept = 0) +
  geom_errorbarh() +
  geom_point() +
  theme_minimal(base_size = 16) +
  theme(plot.title = element_text(hjust = 1, size = 11),
        axis.title = element_text(size = 10),
        axis.text.y = element_blank(),  # Remove y-axis labels
        axis.text = element_text(size = 5),
        plot.caption = element_text(size = 7)) +
  labs(x = "Estimates (95%CIs)",
       y = "Genes",
       title = "Genes Associated with Smoking (Significant)",
       caption = "DOI: https://doi.org/10.1016/j.placenta.2009.12.016")
ggsave("../results/plots/genes_associated_with_smoking_significant1.png")

```

It is clear that all significant genes are divided into 3 clusters:

-   a ≈ 0.22

-   a ≈ 0.15

-   a ≈ -0.11

This can also be concluded when looking at the numeric values for estimates for all significant genes:

```{r}
placental_gene_expr_estimates_p |> 
  filter(is_significant == "yes")
```

Therefore, we selected three genes to represent the three different clusters. Next, we will plot the linear regressions for these genes.

### Linear regression for 3 genes

```{r}
# Looking at the linear regressions for three probe_IDs representing each of the 3 different estimated
lr_ilmn_1651278 <- placental_gene_expr_clean_aug_long |>
  filter(gene == "ilmn_1651278_snip1") |>  
  ggplot(aes(x = status_num, 
             y = log2_expr,
             color = factor(status_num))) +
  geom_point(alpha = 0.6) +  
  geom_smooth(method = "lm", 
              se = TRUE, 
              color = "black") +  
  theme_minimal(base_size = 12) +
  theme(plot.caption = element_text(size = 8)) +
  labs(x = "",
       y = "",
       title = "ILMN_1651278_snip1",
       color = "Smoking Status") +  # Set legend title
  scale_color_manual(values = c("0" = "#F9A1A1", "1" = "#63B8B8"),
                     labels = c("0" = "Non-smoker", "1" = "Smoker"))  # Set legend labels
  

lr_ilmn_1651375 <- placental_gene_expr_clean_aug_long |>
  filter(gene == "ilmn_1651375_bicd2") |>  
  ggplot(aes(x = status_num, 
             y = log2_expr,
             color = factor(status_num))) +
  geom_point(alpha = 0.6) +  
  geom_smooth(method = "lm", 
              se = TRUE, 
              color = "black") +  
  theme_minimal(base_size = 12) +
  theme(plot.caption = element_text(size = 8)) +
  labs(x = "",
    y = "Log2 Expr. level",
    title = "ILMN_1651375_bicd2",
    color = "Smoking Status") +  # Set legend title
  scale_color_manual(values = c("0" = "#F9A1A1", "1" = "#63B8B8"),
                     labels = c("0" = "Non-smoker", "1" = "Smoker"))  # Set legend labels)


lr_ilmn_1651767 <- placental_gene_expr_clean_aug_long |>
  filter(gene == "ilmn_1651767_mkl1") |>  
  ggplot(aes(x = status_num, 
             y = log2_expr,
             color=factor(status_num))) +
  geom_point(alpha = 0.6) +  
  geom_smooth(method = "lm", 
              se = TRUE,
              color = "black") +  
  theme_minimal(base_size = 12) +
  theme(plot.caption = element_text(size = 8)) +
 labs(x = "Status",
    y = "",
    title = "ILMN_1651767_mkl1",
    color = "Smoking Status") +  # Set legend title
  scale_color_manual(values = c("0" = "#F9A1A1", "1" = "#63B8B8"),
                     labels = c("0" = "Non-smoker", "1" = "Smoker"))  # Set legend labels))

  
# Combine the three plots vertically (stacking them)
#linear_regression_plot <- lr_ilmn_1651278 / 
                         #lr_ilmn_1651375 / 
                         #lr_ilmn_1651767 + 
  #plot_layout(ncol = 1) +  # Stacks the plots vertically (1 column)
  #plot_annotation(
    #title = "Linear Regression of Gene Expression Levels",
    #caption = "DOI: https://doi.org/10.1016/j.placenta.2009.12.016"
  #)

# Save the combined plot
#ggsave("../results/plots/linear_regression_plot.png", linear_regression_plot, height = 15, width = 8, 
       #units = "in")


linear_regression_plot <- lr_ilmn_1651278 / lr_ilmn_1651375 / lr_ilmn_1651767 + 
  plot_layout(guides = "collect", 
              heights = c(5, 5, 5),   # Adjust row heights
              design = "A
                          B
                          C") +    # Designates the layout for each plot
  plot_annotation(
    title = "Linear Regression of Gene Expression Levels",
    caption = "DOI: https://doi.org/10.1016/j.placenta.2009.12.016"
  )


linear_regression_plot

ggsave("../results/plots/linear_regression_plot.png", linear_regression_plot, height = 15, width = 8, units = "in")
```

### Forest plot for three genes

```{r}
placental_gene_expr_estimates_p |> 
  filter(gene %in% c("ilmn_1651278_snip1", "ilmn_1651375_bicd2", "ilmn_1651767_mkl1")) |> 
  ggplot(aes(x = estimate,
             y = fct_reorder(gene, estimate),
             xmin = conf.low,
             xmax = conf.high)) +
  geom_vline(xintercept = 0) +
  geom_errorbarh() +
  geom_point() +
  theme_minimal(base_size = 16) +
  theme(plot.title = element_text(hjust = 1, size = 11),
        axis.title = element_text(size = 10),
        axis.text = element_text(size = 5),
        plot.caption = element_text(size = 7)) +
  labs(x = "Estimates (95%CIs)",
       y = "",
       title = "Genes Associated with Smoking",
       caption = "DOI: https://doi.org/10.1016/j.placenta.2009.12.016")
ggsave("../results/plots/genes_associated_with_smoking_3_genes.png")
```

### Forest plot for 30 random significant genes (p\<0.05)

```{r}
placental_gene_expr_estimates_p |> 
  filter(is_significant == "yes") |> 
  sample_n(30) |> 
  ggplot(aes(x = estimate,
             y = fct_reorder(gene, estimate),
             xmin = conf.low,
             xmax = conf.high)) +
  geom_vline(xintercept = 0) +
  geom_errorbarh() +
  geom_point() +
  theme_minimal(base_size = 16) +
  theme(plot.title = element_text(hjust = 1, size = 11),
        axis.title = element_text(size = 10),
        axis.text = element_text(size = 5),
        plot.caption = element_text(size = 7)) +
  labs(x = "Estimates (95%CIs)",
       y = "",
       title = "Genes Associated with Smoking",
       caption = "DOI: https://doi.org/10.1016/j.placenta.2009.12.016")
ggsave("../results/plots/genes_associated_with_smoking_30_genes.png")
```

### Volcano plot

```{r}
volcano_plot <- placental_gene_expr_estimates_p |>
  mutate(lbl = case_when(
    is_significant == "yes" ~ gene,
    is_significant == "no" ~ "")) |> 
  ggplot(aes(x = estimate,
             y = -log10(p.value),
             colour = is_significant,
             label = lbl)) +
  geom_point(size = 1,
             alpha = 0.5) +
  geom_text_repel(size = 4,
                  max.overlaps = 20) +
  geom_hline(yintercept = 0) +
  theme_minimal(base_size = 10) +
  theme(
    plot.title = element_text(hjust = 1),
    plot.subtitle = element_text(hjust = 1),
    legend.position = "none",
    plot.caption = element_text(size = 6)  # Set smaller caption size here
  ) +
  labs(
    x = "Estimates",
    y = "-log10(p)",
    title = "Volcano plot for down/up regulated genes with significance p < 0.05",
    subtitle = "Genes highlighted in turquoise had a significant p-value",
    caption = "DOI: https://doi.org/10.1016/j.placenta.2009.12.016")

volcano_plot
ggsave("../results/plots/volcano_plot_p_value.png")
```
